// src/api/generated/model/aPIEntitiesGroup.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GitLab REST API
 * GitLab REST API used to interact with a GitLab installation.
 * OpenAPI spec version: v4
 */
import type { APIEntitiesCustomAttribute } from './aPIEntitiesCustomAttribute';
import type { APIEntitiesNamespaceRootStorageStatistics } from './aPIEntitiesNamespaceRootStorageStatistics';
import type { APIEntitiesLdapGroupLink } from './aPIEntitiesLdapGroupLink';
import type { APIEntitiesSamlGroupLink } from './aPIEntitiesSamlGroupLink';

export interface APIEntitiesGroup {
  id?: number;
  web_url?: string;
  name?: string;
  path?: string;
  description?: string;
  visibility?: string;
  share_with_group_lock?: boolean;
  require_two_factor_authentication?: boolean;
  two_factor_grace_period?: number;
  project_creation_level?: string;
  auto_devops_enabled?: string;
  subgroup_creation_level?: string;
  emails_disabled?: boolean;
  emails_enabled?: boolean;
  show_diff_preview_in_email?: boolean;
  mentions_disabled?: string;
  lfs_enabled?: boolean;
  archived?: boolean;
  math_rendering_limits_enabled?: boolean;
  lock_math_rendering_limits_enabled?: boolean;
  crm_enabled?: boolean;
  resource_access_token_notify_inherited?: boolean;
  lock_resource_access_token_notify_inherited?: boolean;
  default_branch?: string;
  default_branch_protection?: number;
  default_branch_protection_defaults?: string;
  avatar_url?: string;
  request_access_enabled?: boolean;
  full_name?: string;
  full_path?: string;
  created_at?: string;
  parent_id?: string;
  organization_id?: number;
  shared_runners_setting?: string;
  max_artifacts_size?: number;
  custom_attributes?: APIEntitiesCustomAttribute;
  statistics?: string;
  marked_for_deletion_on?: string;
  root_storage_statistics?: APIEntitiesNamespaceRootStorageStatistics;
  ldap_cn?: string;
  ldap_access?: string;
  ldap_group_links?: APIEntitiesLdapGroupLink;
  saml_group_links?: APIEntitiesSamlGroupLink;
  file_template_project_id?: string;
  wiki_access_level?: string;
  repository_storage?: string;
  /** [Experimental] Indicates whether GitLab Duo Core features are enabled for the group */
  duo_core_features_enabled?: boolean;
  duo_features_enabled?: string;
  lock_duo_features_enabled?: string;
  auto_duo_code_review_enabled?: string;
  web_based_commit_signing_enabled?: string;
  allow_personal_snippets?: string;
  duo_namespace_access_rules?: string;
  built_in_project_templates_enabled?: boolean;
  lock_built_in_project_templates_enabled?: boolean;
}